:recycle: rename the package
