+++ api/KlineAllStock.py
@@ -8,7 +8,7 @@ from utils.dateUtils import get_current_timestamp_millisecond
 
 class KlineAllStock(Resource):
     def get(self):
-        from task import taskManager, sync_kline_day_all
+        from tasks import taskManager, sync_kline_day_all
         taskManager.add_job(func=sync_kline_day_all.run,
                           id=f'sync_kline_all_{get_current_timestamp_millisecond()}'
                           , next_run_time=datetime.now(), executor="default")
