+++ api/Test.py
@@ -1,6 +1,6 @@
 from flask_restful import Resource
 from api.Response_Code import OK
-from task.sync_kline_day_all import run
+from tasks.sync_kline_day_all import run
 
 
 class Test(Resource):
