--- task/__init__.py
@@ -1,3 +1,0 @@
-from task.TaskManager import TaskManager
-
-taskManager = TaskManager()
