+++ tasks/__init__.py
@@ -1,0 +1,3 @@
+from tasks.TaskManager import TaskManager
+
+taskManager = TaskManager()
